refactor class files
